Restructure client project into modules
