+++ packages/client/src/app/auth/services/auth.service.js
@@ -1,4 +1,4 @@
-import { HttpRequestService } from './http-request.service';
+import { HttpRequestService } from '../../core/index.js';
 
 export class AuthService {
   /** @return {AuthService} */
