+++ packages/client/src/app/auth/views/index.js
@@ -1,0 +1,4 @@
+import LoginView from './LoginView.vue';
+import RegisterView from './RegisterView.vue';
+
+export { LoginView, RegisterView };
