+++ packages/client/src/app/challenge/components/challenge-progress/ChallengeCheckBox.vue
