+++ packages/client/src/app/challenge/components/index.js
@@ -1,0 +1,3 @@
+import ChallengeProgress from './challenge-progress/ChallengeProgress.vue';
+
+export { ChallengeProgress };
