+++ packages/client/src/app/challenge/views/ChallengeActiveView.vue
@@ -24,9 +24,9 @@
 </template>
 
 <script>
-import ChallengeProgress from '../components/challenge/progress/ChallengeProgress.vue';
-import data from '../components/challenge/progress/data.json';
-import { useAuthStore } from '../stores';
+import { ChallengeProgress } from '../components/index.js';
+import data from '../data.json';
+import { useAuthStore } from '../../auth/index.js';
 
 /* TODO: send in user and challenge, move logic to backend */
