+++ packages/client/src/app/challenge/views/index.js
@@ -1,0 +1,4 @@
+import ChallengeActiveView from './ChallengeActiveView.vue';
+import ChallengeCreationView from './ChallengeCreationView.vue';
+
+export { ChallengeActiveView, ChallengeCreationView };
