+++ packages/client/src/app/config/index.js
@@ -1,0 +1,1 @@
+export * from './formkit';
