+++ packages/client/src/app/core/index.js
@@ -1,0 +1,2 @@
+export * from './services';
+export * from './views';
