+++ packages/client/src/app/core/services/index.js
@@ -1,0 +1,1 @@
+export * from './http-request.service.js';
