+++ packages/client/src/app/core/views/index.js
@@ -1,0 +1,3 @@
+import HomeView from './HomeView.vue';
+
+export { HomeView };
