+++ packages/client/src/app/index.js
@@ -1,8 +1,4 @@
 import App from './App.vue';
 
 export * from './app-router';
-export * from './components';
-export * from './formkit';
-export * from './stores';
-export * from './views';
 export { App };
