+++ packages/client/src/app/shared/components/index.js
@@ -1,4 +1,4 @@
 import CustomFormKit from './form/CustomFormKit.vue';
-import NavigationBar from './NavigationBar.vue';
+import NavigationBar from './navigation-bar/NavigationBar.vue';
 
 export { CustomFormKit, NavigationBar };
