--- packages/client/src/app/stores/index.js
@@ -1,1 +1,0 @@
-export * from './auth.store';
