+++ packages/client/src/app/user/index.js
@@ -1,0 +1,2 @@
+export * from './models/user.model.js';
+export * from './services/user.service.js';
