--- packages/client/src/app/views/index.js
@@ -1,7 +1,0 @@
-import ChallengeActiveView from './ChallengeActiveView.vue';
-import ChallengeCreationView from './ChallengeCreationView.vue';
-import HomeView from './HomeView.vue';
-import LoginView from './authentication/LoginView.vue';
-import RegisterView from './authentication/RegisterView.vue';
-
-export { ChallengeActiveView, ChallengeCreationView, HomeView, LoginView, RegisterView };
